test(youtube): cover html bootstrap parser
